% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{possible_checks}
\alias{possible_checks}
\title{Possible Constraints to Check}
\format{
A character vector
c("res_type",
"far",
"fl_area",
"fl_area_first",
"fl_area_top",
"footprint",
"height",
"height_eave",
"lot_cov_bldg",
"lot_area",
"parking_enclosed",
"stories",
"unit_0bed",
"unit_1bed",
"unit_2bed",
"unit_3bed",
"unit_4bed",
"unit_density",
"unit_pct_0bed",
"unit_pct_1bed",
"unit_pct_2bed",
"unit_pct_3bed",
"unit_pct_4bed",
"total_units",
"unit_size_avg",
"unit_size",
"bldg_fit",
"overlay")
}
\usage{
possible_checks
}
\description{
This character vector lists all the possible constraints that can be
checked in the \code{zr_run_zoning_checks()} functions
}
\keyword{datasets}
