% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reasons}
\alias{reasons}
\title{Table of zoneR result's False and Maybe reasons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 3 columns.
\describe{
\item{short_reason}{The reason given in the results of \code{zr_run_zoning_checks()}}
\item{false_explain}{What the reason means when the value is FALSE (i.e. when the building is not allowed)}
\item{maybe_explain}{What the reason means when the value is MAYBE (i.e. when the building might be allowed)}
}
}
\source{
Generated internally by zoneR
}
\usage{
reasons
}
\description{
This table describes what each name means when it is listed
as a reason for a MAYBE or FALSE value in the building check.
}
\keyword{datasets}
