% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_check_constraints.R
\name{zr_check_constraints}
\alias{zr_check_constraints}
\title{Is the building allowed base on zoning constraints?}
\usage{
zr_check_constraints(
  vars,
  zoning_req,
  checks = c("far", "fl_area", "fl_area_first", "fl_area_top", "footprint", "height",
    "height_eave", "lot_cov_bldg", "lot_size", "parking_enclosed", "stories",
    "unit_0bed", "unit_1bed", "unit_2bed", "unit_3bed", "unit_4bed", "unit_density",
    "unit_pct_0bed", "unit_pct_1bed", "unit_pct_2bed", "unit_pct_3bed", "unit_pct_4bed",
    "total_units", "unit_size_avg")
)
}
\arguments{
\item{vars}{the result from the \code{zr_get_variables()} function.
If this data frame is supplied, bldg_data, parcel_data, and zoning_data
are not needed.}

\item{zoning_req}{The output of \code{zr_get_zoning_req()}. A data frame with
constraint values}

\item{checks}{A list of all the checks that should take place. The default is
every check possible. Note, if a zoning file doesn't have zoning info for one
of the constraints listed in the checks variable, then it is assumed that
building characteristic is allowed.}
}
\description{
\code{zr_check_constraints()} compares the building characteristics
against many of the main constraints in the \code{.zoning} file. It returns
a one-row data frame with a column for each constraint it checked against.
}
