% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_check_unit.R
\name{zr_check_unit}
\alias{zr_check_unit}
\title{Compare unit sizes and allowed unit sizes}
\usage{
zr_check_unit(district_data, bldg_data, vars)
}
\arguments{
\item{district_data}{one row (representing one district) of a
zoning data frame created from the OZFS *.zoning file}

\item{bldg_data}{either the file path to an OZFS *.bldg file or
a list created from the the *.bldg file using \code{rjson::fromJSON}}

\item{vars}{the result from the \code{get_variables()} function.
If this data frame is supplied, bldg_data, parcel_data, and zoning_data
are not needed.}
}
\value{
Returns TRUE, FALSE, or MAYBE stating whether or not the building would
be allowed in the district based on the unit sizes.
}
\description{
\code{zr_check_unit()} checks to see if the district's zoning
code allows the building based on unit sizes.
}
