% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_get_parcel_dims.R
\name{zr_get_parcel_dims}
\alias{zr_get_parcel_dims}
\title{Isolate dimensional rows from tidyparcel data}
\usage{
zr_get_parcel_dims(parcels_file)
}
\arguments{
\item{parcels_file}{Either a file path to the a parcels geojson file or an sf object with the same data.}
}
\value{
an sf object with only dimensional and centroid data for each parcel
}
\description{
Isolate dimensional rows from tidyparcel data
}
