% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_get_parcel_geo.R
\name{zr_get_parcel_geo}
\alias{zr_get_parcel_geo}
\title{Isolate known side geometry from tidyparcel data}
\usage{
zr_get_parcel_geo(parcels_file)
}
\arguments{
\item{parcels_file}{Either a file path to the a parcels geojson file or an sf object with the same data.}
}
\value{
an sf object with only the side geometries of the parcels with identified sides
}
\description{
Isolate known side geometry from tidyparcel data
}
