% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_merge_overlay_reqs.R
\name{zr_merge_overlay_reqs}
\alias{zr_merge_overlay_reqs}
\title{Merge an overlay and a base district's requirements}
\usage{
zr_merge_overlay_reqs(base_reqs, overlay_reqs, overlay_type)
}
\arguments{
\item{base_reqs}{The data frame representing the zoning requirements for the
base district that was created with \code{zr_get_zoning_req()}}

\item{overlay_reqs}{The data frame representing the zoning requirements for the
overlay district that was created with \code{zr_get_zoning_req()}}

\item{overlay_type}{String stating whether the overlay is of type "restrict", "relax", or "replace"}
}
\value{
One data frame that is a merger of the two input data frames according to
respective overly rules
}
\description{
Takes the overlay district requirements and the base districts requirements, and
creates one data frame of the proper constraint values to be used in the analysis.
}
\examples{
base_reqs <- data.frame(constraint_name = c("lot_area","setback_front"),
                        min_value = I(list(list(0.17), list(5))),
                        max_value = I(list(list(1.5), list(30))))
overlay_reqs <- data.frame(constraint_name = c("lot_area","setback_front"),
                        min_value = I(list(list(0.2), list(10))),
                        max_value = I(list(list(2), list(35))))

zr_merge_overlay_reqs(base_reqs, overlay_reqs, "restrict")

zr_merge_overlay_reqs(base_reqs, overlay_reqs, "relax")

zr_merge_overlay_reqs(base_reqs, overlay_reqs, "replace")

}
