% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_summary_by_muni.R
\name{zr_summary_by_muni}
\alias{zr_summary_by_muni}
\title{Summarize the results of zr_run_zoning_checks() by municipality}
\usage{
zr_summary_by_muni(detailed_result)
}
\arguments{
\item{detailed_result}{The output simple features data frame from
zr_run_zoning_checks()}
}
\value{
A tibble summarizing the number of parcels with TRUE, MAYBE, and FALSE
results by municipality
}
\description{
\code{zr_summary_by_muni()} summarizes results of zr_run_zoning_checks() by municipality
}
\examples{
zoning_file <- zr_example_files("Paradise.zoning")
parcel_file <- zr_example_files("Paradise.parcel")
bldg_file <- zr_example_files("2_fam.bldg")

# only checking against height constraints
just_check_height <- zr_run_zoning_checks(bldg_file = bldg_file,
                                          parcel_files = parcel_file,
                                          zoning_files = zoning_file,
                                          checks = "height")

zr_summary_by_muni(just_check_height)

zr_summary_false(just_check_height)

zr_summary_maybe(just_check_height)


}
