% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_summary_false.R
\name{zr_summary_false}
\alias{zr_summary_false}
\title{Summarize the reasons for maybe results}
\usage{
zr_summary_false(detailed_result, muni = "All")
}
\arguments{
\item{detailed_result}{The output simple features data frame from
zr_run_zoning_checks()}

\item{muni}{The city to summarize data for. Defaults to "All" to aggregate
across all cities include in the results file.}
}
\value{
A tibble summarizing the reasons for maybe values.
}
\description{
\code{zr_summary_by_muni()} summarizes results reasons given for maybe results from
the output of zr_run_zoning_checks()
}
\examples{
zoning_file <- zr_example_files("Paradise.zoning")
parcel_file <- zr_example_files("Paradise.parcel")
bldg_file <- zr_example_files("2_fam.bldg")

# only checking against height constraints
just_check_height <- zr_run_zoning_checks(bldg_file = bldg_file,
                                          parcel_files = parcel_file,
                                          zoning_files = zoning_file,
                                          checks = "height")

zr_summary_by_muni(just_check_height)

zr_summary_false(just_check_height)

zr_summary_maybe(just_check_height)


}
