% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_summary_maybe.R
\name{zr_summary_maybe}
\alias{zr_summary_maybe}
\title{Summarize the reasons for maybe results}
\usage{
zr_summary_maybe(detailed_result)
}
\arguments{
\item{detailed_result}{The output simple features data frame from
zr_run_zoning_checks()}
}
\value{
A tibble summarizing the reasons for maybe values.
}
\description{
\code{zr_summary_by_muni()} summarizes results reasons given for maybe results from
the output of zr_run_zoning_checks()
}
