% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_check_res_type.R
\name{zr_check_res_type}
\alias{zr_check_res_type}
\title{Is the building allowed base on use_type?}
\usage{
zr_check_res_type(vars, district_data)
}
\arguments{
\item{vars}{the result from the \code{zr_get_variables()} function.
If this data frame is supplied, bldg_data, parcel_data, and zoning_data
are not needed.}

\item{district_data}{one row (representing one district) of a
zoning data frame created from the OZFS *.zoning file}
}
\value{
TRUE or FALSE
}
\description{
Campares the building use type with the permitted use types in the zoning
code and returns TRUE or FALSE
}
