% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_find_district_idx.R
\name{zr_find_district_idx}
\alias{zr_find_district_idx}
\title{Find corresponding zoning district}
\usage{
zr_find_district_idx(
  parcels_centroids_sf,
  tidyzoning,
  idx_col_name = "zoning_id"
)
}
\arguments{
\item{parcels_centroids_sf}{An sf object with the parcel centroids}

\item{tidyzoning}{A simple features object with a row for each zoning district and columns holding the geojson formatted zoning requirements of each district.}

\item{idx_col_name}{The name of the new column that will be added with the district index}
}
\value{
The same parcels centroids object that was input but with an additional
column stating the row number of the corresponding tidyzoning district
}
\description{
\code{zr_find_district_idx()} uses \code{\link[sf:st_join]{sf::st_join()}} to adds a new column to the parcels data that
contains the tidyzoning row index of the district the parcel is in
}
