% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_get_buildable_area.R
\name{zr_get_buildable_area}
\alias{zr_get_buildable_area}
\title{Create a parcel's buildable area}
\usage{
zr_get_buildable_area(parcel_with_setbacks, crs = NULL)
}
\arguments{
\item{parcel_with_setbacks}{A parcel_geo object that has setback
information added. parcel_with_setbacks is the output of the
\code{\link[=zr_add_setbacks]{zr_add_setbacks()}} function.}

\item{crs}{The projected Coordinate Reference System for the study area.
Must be in meters. If NULL, crs is calculated using \code{zr_get_crs()}.}
}
\value{
Returns a polygon representing the buildable area of the parcel.
}
\description{
\code{zr_get_buildable_area()} takes a parcel with setback information and
produces a polygon representing the buildable area of the parcel.
}
