% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_run_zoning_checks.R
\name{zr_run_zoning_checks}
\alias{zr_run_zoning_checks}
\title{Find where a building is allowed to be built}
\usage{
zr_run_zoning_checks(
  bldg_file,
  parcel_files,
  zoning_files,
  detailed_check = FALSE,
  print_checkpoints = TRUE,
  checks = possible_checks,
  save_to = NULL
)
}
\arguments{
\item{bldg_file}{The path to the OZFS *.bldg}

\item{parcel_files}{The path to an OZFS \code{.parcel} file, or the path to a
folder containing OZFS \code{.parcel} files}

\item{zoning_files}{The path to an OZFS \code{.zoning} file, or the path to a
folder containing OZFS \code{.zoning} files}

\item{detailed_check}{When TRUE, every parcel passes through each
check no matter the result, and it takes more time. When FALSE,
subsequent checks are skipped as soon as one check reads FALSE}

\item{print_checkpoints}{When TRUE, runtimes and other info will be
printed at certain points throughout the function.}

\item{checks}{A list of all the checks that should take place. The default is
to check for every constraint possible in the OZFS. These constraints can found
in the package data \code{possible_checks}. Note, if a zoning file doesn't have zoning
info for one of the constraints listed in the checks variable, then it is
assumed that building characteristic is allowed.}

\item{save_to}{The path or folder directory where you want to save the results
as a geojson file.}
}
\value{
A simple features data frame with geometry for the centroid of each
parcel and columns to show info on allowance of the building.
}
\description{
\code{zr_run_zoning_checks()} checks the building information against all the
zoning constraints to see which parcels will allow the building.
}
\section{Zoning Checks}{

The checks that can take place in \code{zr_run_zoning_checks()} are included in
the table below. Most checks are not required and only checked if specified
by the user in the \code{checks} argument. All optional checks possible are
stored in the \code{possible_checks} data which is the default value of the \code{checks}
argument. A few checks are built in to the function and, therefore, not a part of the
\code{checks} argument.

The resulting data frame of \code{zr_run_zoning_checks()} will have a "reason" column with
the names of any checks that caused a FALSE or a MAYBE building allowance.
The table below specifies what each check means when it is part of the "reason"
column.\tabular{lll}{
   Possible Checks \tab Type \tab Reason Column Usage \cr
   PD_dist \tab built in \tab FALSE returned because parcel is in a planned development district \cr
   PD_overlay \tab built in \tab FALSE returned because parcel is in a planned development overlay district \cr
   side_lbl \tab built in \tab MAYBE returned because the parcel sides were not labeled and check_fit was not run \cr
   res_type \tab user input \tab Residential type check returned FALSE or MAYBE \cr
   far \tab user input \tab Floor Area Ratio check returned FALSE or MAYBE \cr
   fl_area \tab user input \tab Floor area check returned FALSE or MAYBE \cr
   fl_area_first \tab user input \tab Area of first floor check returned FALSE or MAYBE \cr
   fl_area_top \tab user input \tab Area of top floor check returned FALSE or MAYBE \cr
   footprint \tab user input \tab Footprint check returned FALSE or MAYBE \cr
   height \tab user input \tab Height check returned FALSE or MAYBE \cr
   height_eave \tab user input \tab Height eave check returned FALSE or MAYBE \cr
   lot_cov_bldg \tab user input \tab Lot coverage check returned FALSE or MAYBE \cr
   lot_area \tab user input \tab Lot area check returned FALSE or MAYBE \cr
   parking_enclosed \tab user input \tab Enclosed parking check returned FALSE or MAYBE \cr
   stories \tab user input \tab Stories check returned FALSE or MAYBE \cr
   unit_0bed \tab user input \tab 0-bed unit quantity check returned FALSE or MAYBE \cr
   unit_1bed \tab user input \tab 1-bed unit quantity check returned FALSE or MAYBE \cr
   unit_2bed \tab user input \tab 2-bed unit quantity check returned FALSE or MAYBE \cr
   unit_3bed \tab user input \tab 3-bed unit quantity check returned FALSE or MAYBE \cr
   unit_4bed \tab user input \tab 4plus-bed unit quantity check returned FALSE or MAYBE \cr
   unit_density \tab user input \tab Unit density check returned FALSE or MAYBE \cr
   unit_pct_0bed \tab user input \tab Percent 0-bed units check returned FALSE or MAYBE \cr
   unit_pct_1bed \tab user input \tab Percent 1-bed units check returned FALSE or MAYBE \cr
   unit_pct_2bed \tab user input \tab Percent 2-bed units check returned FALSE or MAYBE \cr
   unit_pct_3bed \tab user input \tab Percent 3-bed units check returned FALSE or MAYBE \cr
   unit_pct_4bed \tab user input \tab Percent 4plus-bed units check returned FALSE or MAYBE \cr
   total_units \tab user input \tab Total units check returned FALSE or MAYBE \cr
   unit_size_avg \tab user input \tab Average unit size check returned FALSE or MAYBE \cr
   unit_size \tab user input \tab Unit size check returned FALSE or MAYBE \cr
   bldg_fit \tab user input \tab Building fit check returned FALSE or MAYBE \cr
   overlay \tab user input \tab MAYBE returned because parcel is in an overlay district \cr
}
}

