% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_run_zoning_checks.R
\name{zr_run_zoning_checks}
\alias{zr_run_zoning_checks}
\title{Find where a building is allowed to be built}
\usage{
zr_run_zoning_checks(
  bldg_file,
  parcel_files,
  zoning_files,
  detailed_check = FALSE,
  print_checkpoints = TRUE,
  checks = possible_checks,
  save_to = NULL
)
}
\arguments{
\item{bldg_file}{The path to the OZFS *.bldg}

\item{parcel_files}{The path, or list of paths, to the OZFS *.parcel file or files}

\item{zoning_files}{The path, or list of paths, to the OZFS *.zoning file or files}

\item{detailed_check}{When TRUE, every parcel passes through each
check no matter the result, and it takes more time. When FALSE,
subsequent checks are skipped as soon as one check reads FALSE}

\item{print_checkpoints}{When TRUE, runtimes and other info will be
printed at certain points throughout the function.}

\item{checks}{A list of all the checks that should take place. The default is
to check for every constraint possible in the OZFS. These constraints can found
in the package data \code{possible_checks}. Note, if a zoning file doesn't have zoning
info for one of the constraints listed in the checks variable, then it is
assumed that building characteristic is allowed.}

\item{save_to}{The path or folder directory where you want to save the results
as a geojson file.}
}
\value{
A simple features data frame with geometry for the centroid of each
parcel and columns to show info on allowance of the building.
}
\description{
\code{zr_run_zoning_checks()} checks the building information against all the
zoning constraints to see which parcels will allow the building.
}
