% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_run_zoning_checks.R
\name{zr_run_zoning_checks}
\alias{zr_run_zoning_checks}
\title{Find where a building is allowed to be built}
\usage{
zr_run_zoning_checks(
  bldg_file,
  parcel_files,
  zoning_files,
  detailed_check = FALSE,
  print_checkpoints = TRUE,
  checks = possible_checks,
  save_to = NULL
)
}
\arguments{
\item{bldg_file}{The path to the OZFS *.bldg}

\item{parcel_files}{The path to an OZFS \code{.parcel} file, or the path to a
folder containing OZFS \code{.parcel} files}

\item{zoning_files}{The path to an OZFS \code{.zoning} file, or the path to a
folder containing OZFS \code{.zoning} files}

\item{detailed_check}{When TRUE, every parcel passes through each
check no matter the result, and it takes more time. When FALSE,
subsequent checks are skipped as soon as one check reads FALSE}

\item{print_checkpoints}{When TRUE, runtimes and other info will be
printed at certain points throughout the function.}

\item{checks}{A list of all the checks that should take place. The default is
to check for every constraint possible in the OZFS. These constraints can found
in the package data \code{possible_checks}. Note, if a zoning file doesn't have zoning
info for one of the constraints listed in the checks variable, then it is
assumed that building characteristic is allowed.}

\item{save_to}{The path or folder directory where you want to save the results
as a geojson file.}
}
\value{
A simple features data frame with geometry for the centroid of each
parcel and columns to show info on allowance of the building.
}
\description{
\code{zr_run_zoning_checks()} checks the building information against all the
zoning constraints to see which parcels will allow the building.
}
\section{Zoning Checks}{

The checks that can take place in \code{zr_run_zoning_checks()} are included in
the table below. Most checks are not required and only checked if specified
by the user in the \code{checks} argument. All optional checks possible are
stored in the \code{possible_checks} data which is the default value of the \code{checks}
argument. A few checks are built in to the function and, therefore, not a part of the
\code{checks} argument.

The resulting data frame of \code{zr_run_zoning_checks()} will have a "reason" column with
the names of any checks that caused a FALSE or a MAYBE building allowance. For
more details on what the reasons column means, see the package data \code{reasons}.
}

\examples{
zoning_file <- zr_example_files("Paradise.zoning")
parcel_file <- zr_example_files("Paradise.parcel")
bldg_file <- zr_example_files("2_fam.bldg")

# only checking against height constraints for speed
just_check_height <- zr_run_zoning_checks(bldg_file = bldg_file,
                                          parcel_files = parcel_file,
                                          zoning_files = zoning_file,
                                          checks = "height")
}
