% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_run_zoning_checks.R
\name{zr_run_zoning_checks}
\alias{zr_run_zoning_checks}
\title{Find where a building is allowed to be built}
\usage{
zr_run_zoning_checks(
  bldg_file,
  parcels_file,
  zoning_file,
  detailed_check = TRUE,
  print_checkpoints = TRUE,
  checks = c("res_type", "far", "fl_area", "fl_area_first", "fl_area_top", "footprint",
    "height", "height_eave", "lot_cov_bldg", "lot_size", "parking_enclosed", "stories",
    "unit_0bed", "unit_1bed", "unit_2bed", "unit_3bed", "unit_4bed", "unit_density",
    "unit_pct_0bed", "unit_pct_1bed", "unit_pct_2bed", "unit_pct_3bed", "unit_pct_4bed",
    "total_units", "unit_size_avg", "unit_size", "bldg_fit", "overlay")
)
}
\arguments{
\item{bldg_file}{The path to the OZFS *.bldg file}

\item{parcels_file}{The path to the OZFS *.parcel file}

\item{zoning_file}{The path to the OZFS *.zoning}

\item{detailed_check}{When TRUE, every parcel passes through each
check no matter the result, and it take more time. When FALSE,
subsequent checks are skipped as soon as one check reads FALSE}

\item{print_checkpoints}{When TRUE, runtimes and other info will be
printed at certain points throughout the function.}

\item{checks}{A list of all the checks that should take place. The default is
every check possible. Note, if a zoning file doesn't have zoning info for one
of the constraints listed in the checks variable, then it is assumed that
building characteristic is allowed.}
}
\value{
a simple features data frame with the centroid of each parcel with a column
stating building allowance on the parcel and a column stating the reason
why certain parcels don't allow the building.
}
\description{
\code{zr_run_zoning_checks()} checks the building information against all the
zoning constraints to see which parcels will allow the building.
}
