% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr_summary_by_muni.R
\name{zr_summary_by_muni}
\alias{zr_summary_by_muni}
\title{Summarize the results of zr_run_zoning_checks() by municipality}
\usage{
zr_summary_by_muni(detailed_result)
}
\arguments{
\item{detailed_result}{The output simple features data frame from
zr_run_zoning_checks()}
}
\value{
A tibble summarizing the number of parcels with TRUE, MAYBE, and FALSE
results by municipality
}
\description{
\code{zr_summary_by_muni()} summarizes results of zr_run_zoning_checks() by municipality
}
